requirements.txt
google-genai>=1.0.0
fish-audio-sdk>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.115.0
python-multipart>=0.0.7
uvicorn>=0.32.0
sse-starlette>=2.1.0
